feat: add stock finder component and spinner.
Stock finder component is a search input that will search for a stock
symbol and return a list of results. The user can then select a stock
symbol from the list and the stock price component will be updated with
the new stock symbol based in a custom event.

Spinner component is a simple spinner that will be displayed.